labs/build-production-ready-ai-solutions-with-claude/requirements.txt
boto3
uv
strands-agents
strands-agents-tools
weasyprint
bedrock-agentcore>=0.1.2
bedrock-agentcore-starter-toolkit==0.1.5
aws-opentelemetry-distro
ddgs
aws-opentelemetry-distro~=0.10.1
pyyaml
nest-asyncio